Replace public key with X-only public key
